+++ app/src/main/java/woowacourse/shopping/data/DummyProduct.kt
@@ -1,4 +1,6 @@
-package woowacourse.shopping
+package woowacourse.shopping.data
+
+import woowacourse.shopping.model.Product
 
 val dummyProducts: List<Product> =
     listOf(
