+++ app/src/main/java/woowacourse/shopping/feature/main/MainActivity.kt
@@ -1,8 +1,14 @@
-package woowacourse.shopping
+package woowacourse.shopping.feature.main
 
 import android.os.Bundle
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
+import woowacourse.shopping.feature.main.adapter.ProductAdapter
+import woowacourse.shopping.feature.detail.ProductDetailActivity
+import woowacourse.shopping.data.ProductRepository
+import woowacourse.shopping.data.ProductRepositoryImpl
+import woowacourse.shopping.viewmodel.ProductsViewModel
+import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityMainBinding
 
 class MainActivity : AppCompatActivity() {
