+++ app/src/main/java/woowacourse/shopping/viewmodel/ProductsViewModel.kt
@@ -1,8 +1,10 @@
-package woowacourse.shopping
+package woowacourse.shopping.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import woowacourse.shopping.data.ProductRepository
+import woowacourse.shopping.model.Product
 
 class ProductsViewModel : ViewModel() {
     private val _products = MutableLiveData<List<Product>>()
