+++ app/src/main/res/layout/item_product.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="product"
-            type="woowacourse.shopping.Product" />
+            type="woowacourse.shopping.model.Product" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
